requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
moviepy>=1.0.3,<2.0
Pillow>=10.0.0
numpy>=1.26.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
schedule>=1.2.0
